Make generic entity constructor
