+++ gallery/src/data/provide_hass.js
@@ -1,130 +1,8 @@
 import fireEvent from '../../../src/common/dom/fire_event.js';
 
 const ensureArray = val => (Array.isArray(val) ? val : [val]);
-const now = () => new Date().toISOString();
-const randomTime = () =>
-  new Date(new Date().getTime() - (Math.random() * 80 * 60 * 1000)).toISOString();
 
-export class Entity {
-  constructor(domain, objectId, state, baseAttributes = {}) {
-    this.domain = domain;
-    this.objectId = objectId;
-    this.entityId = `${domain}.${objectId}`;
-    this.lastChanged = randomTime();
-    this.lastUpdated = randomTime();
-    this.state = state;
-    // These are the attributes that we always write to the state machine
-    this.baseAttributes = baseAttributes;
-    this.attributes = baseAttributes;
-  }
-
-  // eslint-disable-next-line
-  async handleService(domain, service, data) {
-  }
-
-  update(state, attributes = {}) {
-    this.state = state;
-    this.lastUpdated = now();
-    this.lastChanged = state === this.state ? this.lastChanged : this.lastUpdated;
-    this.attributes = Object.assign({}, this.baseAttributes, attributes);
-
-    console.log('update', this.entityId, this);
-
-    this.hass.updateStates({
-      [this.entityId]: this.toState()
-    });
-  }
-
-  toState() {
-    return {
-      entity_id: this.entityId,
-      state: this.state,
-      attributes: this.attributes,
-      last_changed: this.lastChanged,
-      last_updated: this.lastUpdated,
-    };
-  }
-}
-
-export class LightEntity extends Entity {
-  constructor(objectId, state, baseAttributes) {
-    super('light', objectId, state ? 'on' : 'off', baseAttributes);
-  }
-
-  async handleService(domain, service, data) {
-    if (domain !== this.domain) return;
-
-    if (service === 'turn_on') {
-      // eslint-disable-next-line
-      const { brightness, hs_color } = data;
-      this.update('on', Object.assign(this.attributes, {
-        brightness,
-        hs_color,
-      }));
-    } else if (service === 'turn_off') {
-      this.update('off');
-    } else if (service === 'toggle') {
-      if (this.state === 'on') {
-        this.handleService(domain, 'turn_off', data);
-      } else {
-        this.handleService(domain, 'turn_on', data);
-      }
-    }
-  }
-}
-
-export class LockEntity extends Entity {
-  constructor(objectId, state, baseAttributes) {
-    super('lock', objectId, state ? 'locked' : 'unlocked', baseAttributes);
-  }
-
-  // eslint-disable-next-line
-  async handleService(domain, service, data) {
-    if (domain !== this.domain) return;
-
-    if (service === 'lock') {
-      this.update('locked');
-    } else if (service === 'unlock') {
-      this.update('unlocked');
-    }
-  }
-}
-
-export class CoverEntity extends Entity {
-  constructor(objectId, state, baseAttributes) {
-    super('cover', objectId, state, baseAttributes);
-  }
-
-  // eslint-disable-next-line
-  async handleService(domain, service, data) {
-    if (domain !== this.domain) return;
-
-    if (service === 'open_cover') {
-      this.update('open');
-    } else if (service === 'close_cover') {
-      this.update('closing');
-    }
-  }
-}
-
-export class GroupEntity extends Entity {
-  constructor(objectId, state, baseAttributes) {
-    super('group', objectId, state, baseAttributes);
-  }
-
-  async handleService(domain, service, data) {
-    if (!['group', 'homeassistant'].includes(domain)) return;
-
-    await Promise.all(this.attributes.entity_id.map((ent) => {
-      const entity = this.hass.mockEntities[ent];
-      return entity.handleService(entity.domain, service, data);
-    }));
-
-    this.update(service === 'turn_on' ? 'on' : 'off');
-  }
-}
-
-export function provideHass(elements, { initialStates = {} } = {}) {
+export default (elements, { initialStates = {} } = {}) => {
   elements = ensureArray(elements);
 
   const wsCommands = {};
