+++ gallery/src/demos/demo-hui-entities-card.js
@@ -1,42 +1,36 @@
 import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
-import {
-  provideHass,
-  Entity,
-  CoverEntity,
-  GroupEntity,
-  LightEntity,
-  LockEntity,
-} from '../data/provide_hass.js';
+import getEntity from '../data/entity.js';
+import provideHass from '../data/provide_hass.js';
 import '../components/demo-cards.js';
 
 const ENTITIES = [
-  new LightEntity('bed_light', true, {
+  getEntity('light', 'bed_light', true, {
     friendly_name: 'Bed Light'
   }),
-  new GroupEntity('kitchen', 'on', {
+  getEntity('group', 'kitchen', 'on', {
     entity_id: [
       'light.bed_light',
     ],
     order: 8,
     friendly_name: 'Kitchen'
   }),
-  new LockEntity('kitchen_door', true, {
+  getEntity('lock', 'kitchen_door', true, {
     friendly_name: 'Kitchen Door'
   }),
-  new CoverEntity('kitchen_window', 'open', {
+  getEntity('cover', 'kitchen_window', 'open', {
     friendly_name: 'Kitchen Window',
     supported_features: 11
   }),
-  new Entity('scene', 'romantic_lights', 'scening', {
+  getEntity('scene', 'romantic_lights', 'scening', {
     entity_id: [
       'light.bed_light',
       'light.ceiling_lights'
     ],
     friendly_name: 'Romantic lights'
   }),
-  new Entity('device_tracker', 'demo_paulus', 'home', {
+  getEntity('device_tracker', 'demo_paulus', 'home', {
     source_type: 'gps',
     latitude: 32.877105,
     longitude: 117.232185,
